Split view type definition file
